src/test/java/com/weather/application/controller/HealthCheckControllerTest.java
package com.weather.application.controller;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.web.client.TestRestTemplate;
import org.springframework.boot.test.web.server.LocalServerPort;
import org.springframework.test.context.TestPropertySource;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@TestPropertySource(locations="classpath:application-test.properties")
public class HealthCheckControllerTest {
    @Autowired
    private HealthCheckController controller;
    @LocalServerPort
    private int port;
    @Autowired
    private TestRestTemplate restTemplate;

    @Test
    void contextLoads() throws Exception {
        assertThat(controller).isNotNull();
    }

    @Test
    void healthCheckTest() throws Exception {

        assertThat(this.restTemplate.withBasicAuth("testUser", "testPassword").getForObject("http://localhost:" + port + "/healthCheck",
                String.class)).contains("Welcome to Weather App!");
    }
}